% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regression.R
\name{betahat_normaleqn}
\alias{betahat_normaleqn}
\title{Compute the normal equations solution to least-squares
\eqn{ (X^T X)^{-1} X^T y }}
\usage{
betahat_normaleqn(X, y)
}
\arguments{
\item{X}{a matrix of covariates (n by p)}

\item{y}{a vector of outcomes (n)}
}
\value{
the normal equations least-squares \eqn{ \hat \beta }
}
\description{
Compute the normal equations solution to least-squares
\eqn{ (X^T X)^{-1} X^T y }
}

