% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regression.R
\name{J}
\alias{J}
\title{Compute the linear regression loss function
\eqn{ J(\beta) = \frac{1}{2m} \sum_{i=1}^N (\hat y_i(\beta) - y_i) }}
\usage{
J(X, y, beta)
}
\arguments{
\item{X}{a matrix of covariates (n by p)}

\item{y}{a vector of outcomes (n)}

\item{beta}{a vector of coefficients (p)}
}
\value{
the loss function \eqn{J(\beta)} evaluated on \code{X} and \code{y}.
}
\description{
Compute the linear regression loss function
\eqn{ J(\beta) = \frac{1}{2m} \sum_{i=1}^N (\hat y_i(\beta) - y_i) }
}

